Move fixture to conftest
